+++ package.json
@@ -39,6 +39,8 @@
     "jest": "^27.5.1",
     "lint-staged": "^12.1.7",
     "mini-css-extract-plugin": "^2.6.0",
+    "nodemon": "^2.0.16",
+    "null-loader": "^4.0.1",
     "postcss-loader": "^6.2.1",
     "postcss-scss": "^4.0.3",
     "prettier": "^2.6.1",
@@ -53,10 +55,12 @@
     "stylelint-scss": "^4.2.0",
     "ts-jest": "^27.1.4",
     "ts-loader": "^9.2.8",
+    "tsconfig-paths-webpack-plugin": "^3.5.2",
     "typescript": "^4.6.3",
     "webpack": "^5.70.0",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
+    "webpack-dev-server": "^4.7.4",
+    "webpack-node-externals": "^3.0.0"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.8.1",
@@ -65,6 +69,7 @@
     "express": "^4.18.1",
     "husky": "^7.0.4",
     "normalize.css": "^8.0.1",
+    "npm-run-all": "^4.1.5",
     "path": "^0.12.7",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
@@ -82,7 +87,10 @@
   },
   "scripts": {
     "ssr": "webpack --mode development && node ./packages/server/index.js",
-    "start": "webpack serve --mode development",
+    "start:webpack": "webpack --mode=development --watch",
+    "start:server": "nodemon ./packages/server/index.js --watch dist/server.js",
+    "start": "NODE_ENV=development npm-run-all --print-label --parallel start:*",
+    "devserv": "webpack serve --mode development",
     "build": "webpack --mode production",
     "lint:style": "yarn stylelint './packages/client/src/**/*.{scss,sass}'",
     "lint": "yarn eslint ./packages/client/src --ext .js,.jsx,.ts,.tsx",
