+++ packages/client/src/pages/Home/Home.tsx
@@ -1,11 +1,11 @@
 import React, {ReactElement} from 'react';
 import '@/assets/images/motypest1-min.png';
-import ReactMarkdown from 'react-markdown';
+// import ReactMarkdown from 'react-markdown';
 
 import {Link} from 'react-router-dom';
 import {useAppSelector} from '@/utils';
 import Button from '../../components/Button/Button';
-import Readme from '../../../../../README.md';
+// import Readme from '../../../../../README.md';
 import {ROUTES} from '../consts';
 
 import style from './home.module.scss';
@@ -15,9 +15,7 @@ function Home(): ReactElement {
 
   return (
     <div className={style.wrapper}>
-      <div className={style.content}>
-        <ReactMarkdown>{Readme}</ReactMarkdown>
-      </div>
+      <div className={style.content}>{/* <ReactMarkdown>{Readme}</ReactMarkdown> */}</div>
       <div className={style.controls}>
         {!isAuth ? (
           <Link to={ROUTES.AUTH}>
@@ -29,7 +27,6 @@ function Home(): ReactElement {
           </Link>
         )}
       </div>
-
     </div>
   );
 }
